A quick update to deal with the new website #5
Now has a JSON lib dependency.

This version will stop at either the previous stop date, or 500
pages. Because of gaps in numbering, and because of reverse ordering
in pages, however, it will *always* fetch those 500 metadata pages.
It would be nice to be smarter/nicer about this, but it's hard to be
sure we've hit contiguous old data (as a previous run may have been
interrupted, and we might only have gotten back as far as the start of
that run).